+++ sneakctl_server/core/systemd.py
@@ -47,8 +47,8 @@ class Systemd:
                 count += 1
 
             msg.update({'name': name})
-            print(msg)
             status.append(msg)
+            self.manager.ResetFailed()  # resets the "failed" state of all units... kind of a flush
 
         return {'services': status, 'successful': count, 'failed': len(names) - count}